Pipe/Validator related refactoring and code-cleanup (#7008)
