+++ core/src/main/java/org/frankframework/pipes/XmlValidator.java
@@ -18,7 +18,6 @@ package org.frankframework.pipes;
 import java.io.IOException;
 import java.net.URL;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
@@ -29,16 +28,16 @@ import java.util.StringTokenizer;
 import javax.xml.transform.TransformerConfigurationException;
 import javax.xml.validation.ValidatorHandler;
 
-import jakarta.annotation.Nonnull;
+import lombok.Getter;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.xerces.xs.XSModel;
 import org.frankframework.configuration.ConfigurationException;
 import org.frankframework.configuration.ConfigurationWarnings;
 import org.frankframework.configuration.HasSpecialDefaultValues;
+import org.frankframework.core.AbstractResponseValidatorWrapper;
 import org.frankframework.core.IValidator;
 import org.frankframework.core.IXmlValidator;
 import org.frankframework.core.PipeForward;
-import org.frankframework.core.PipeLine;
 import org.frankframework.core.PipeLineSession;
 import org.frankframework.core.PipeRunException;
 import org.frankframework.core.PipeRunResult;
@@ -50,7 +49,6 @@ import org.frankframework.soap.SoapVersion;
 import org.frankframework.stream.Message;
 import org.frankframework.util.ClassLoaderUtils;
 import org.frankframework.util.ClassUtils;
-import org.frankframework.util.Locker;
 import org.frankframework.util.SpringUtils;
 import org.frankframework.util.TransformerPool;
 import org.frankframework.util.TransformerPool.OutputType;
@@ -70,11 +68,7 @@ import org.frankframework.validation.XmlValidatorException;
 import org.frankframework.validation.xsd.ResourceXsd;
 import org.frankframework.xml.RootElementToSessionKeyFilter;
 import org.springframework.beans.factory.InitializingBean;
-import org.springframework.context.ApplicationContext;
 import org.xml.sax.helpers.XMLFilterImpl;
-
-import lombok.Getter;
-import lombok.Setter;
 
 
 /**
@@ -425,25 +419,12 @@ public class XmlValidator extends ValidatorBase implements SchemasProvider, HasS
 	}
 
 	/**
-	 * Wrapper for the response validator. It has its own name and forwards, but delegates the actual work to the original validator.
-	 * It overrides the stop and start method to prevent the original validator from being started and stopped.
+	 * Wrapper for the XML response validator.
 	 */
-	public static class ResponseValidatorWrapper implements IXmlValidator {
-
-		private @Getter @Setter String name;
-
-		private final Map<String, PipeForward> forwards = new HashMap<>();
-
-		protected XmlValidator owner;
+	public static class ResponseValidatorWrapper extends AbstractResponseValidatorWrapper<XmlValidator> implements IXmlValidator {
 
 		public ResponseValidatorWrapper(XmlValidator owner) {
-			super();
-			this.owner = owner;
-			name = "ResponseValidator of " + owner.getName();
-		}
-
-		@Override
-		public void configure() throws ConfigurationException {
+			super(owner);
 		}
 
 		@Override
@@ -452,41 +433,8 @@ public class XmlValidator extends ValidatorBase implements SchemasProvider, HasS
 		}
 
 		@Override
-		public PipeRunResult doPipe(Message message, PipeLineSession session) throws PipeRunException {
-			return owner.doPipe(message, session, true, null);
-		}
-
-		@Override
-		public PipeRunResult validate(Message message, PipeLineSession session, String messageRoot) throws PipeRunException {
-			return owner.doPipe(message, session, true, messageRoot);
-		}
-
-		@Override
 		public String getMessageRoot() {
 			return owner.getResponseRoot();
-		}
-
-		@Override
-		public int getMaxThreads() {
-			return 0;
-		}
-
-		@Override
-		public Map<String, PipeForward> getForwards() {
-			return forwards;
-		}
-
-		@Override
-		public void registerForward(PipeForward forward) {
-			forwards.put(forward.getName(), forward);
-		}
-
-		@Override
-		public void start() throws PipeStartException {
-		}
-
-		@Override
-		public void stop() {
 		}
 
 		@Override
@@ -507,196 +455,6 @@ public class XmlValidator extends ValidatorBase implements SchemasProvider, HasS
 		@Override
 		public String getDocumentation() {
 			return null;
-		}
-
-		@Override
-		public ApplicationContext getApplicationContext() {
-			return owner.getApplicationContext();
-		}
-
-		@Override
-		public ClassLoader getConfigurationClassLoader() {
-			return owner.getConfigurationClassLoader();
-		}
-
-		@Override
-		public void setApplicationContext(@Nonnull ApplicationContext applicationContext) {
-			//Can ignore this as it's not set through Spring
-		}
-
-		@Override
-		public boolean consumesSessionVariable(String sessionKey) {
-			return owner.consumesSessionVariable(sessionKey);
-		}
-
-		@Override
-		public void setPipeLine(PipeLine pipeline) {
-			owner.setPipeLine(pipeline);
-		}
-
-		@Override
-		public void setGetInputFromSessionKey(String string) {
-			owner.setGetInputFromSessionKey(string);
-		}
-
-		@Override
-		public String getGetInputFromSessionKey() {
-			return owner.getGetInputFromSessionKey();
-		}
-
-		@Override
-		public void setGetInputFromFixedValue(String string) {
-			owner.setGetInputFromFixedValue(string);
-		}
-
-		@Override
-		public String getGetInputFromFixedValue() {
-			return owner.getGetInputFromFixedValue();
-		}
-
-		@Override
-		public void setEmptyInputReplacement(String string) {
-			owner.setEmptyInputReplacement(string);
-		}
-
-		@Override
-		public String getEmptyInputReplacement() {
-			return owner.getEmptyInputReplacement();
-		}
-
-		@Override
-		public void setPreserveInput(boolean preserveInput) {
-			owner.setPreserveInput(preserveInput);
-		}
-
-		@Override
-		public boolean isPreserveInput() {
-			return owner.isPreserveInput();
-		}
-
-		@Override
-		public void setStoreResultInSessionKey(String string) {
-			owner.setStoreResultInSessionKey(string);
-		}
-
-		@Override
-		public String getStoreResultInSessionKey() {
-			return owner.getStoreResultInSessionKey();
-		}
-
-		@Override
-		public void setChompCharSize(String string) {
-			owner.setChompCharSize(string);
-		}
-
-		@Override
-		public String getChompCharSize() {
-			return owner.getChompCharSize();
-		}
-
-		@Override
-		public void setElementToMove(String string) {
-			owner.setElementToMove(string);
-		}
-
-		@Override
-		public String getElementToMove() {
-			return owner.getElementToMove();
-		}
-
-		@Override
-		public void setElementToMoveSessionKey(String string) {
-			owner.setElementToMoveSessionKey(string);
-		}
-
-		@Override
-		public String getElementToMoveSessionKey() {
-			return owner.getElementToMoveSessionKey();
-		}
-
-		@Override
-		public void setElementToMoveChain(String string) {
-			owner.setElementToMoveChain(string);
-		}
-
-		@Override
-		public String getElementToMoveChain() {
-			return owner.getElementToMoveChain();
-		}
-
-		@Override
-		public void setRemoveCompactMsgNamespaces(boolean b) {
-			owner.setRemoveCompactMsgNamespaces(b);
-		}
-
-		@Override
-		public boolean isRemoveCompactMsgNamespaces() {
-			return owner.isRemoveCompactMsgNamespaces();
-		}
-
-		@Override
-		public void setRestoreMovedElements(boolean restoreMovedElements) {
-			owner.setRestoreMovedElements(restoreMovedElements);
-		}
-
-		@Override
-		public boolean isRestoreMovedElements() {
-			return owner.isRestoreMovedElements();
-		}
-
-		@Override
-		public void setDurationThreshold(long maxDuration) {
-			owner.setDurationThreshold(maxDuration);
-		}
-
-		@Override
-		public long getDurationThreshold() {
-			return owner.getDurationThreshold();
-		}
-
-		@Override
-		public void setLocker(Locker locker) {
-			owner.setLocker(locker);
-		}
-
-		@Override
-		public Locker getLocker() {
-			return owner.getLocker();
-		}
-
-		@Override
-		public void setWriteToSecLog(boolean b) {
-			owner.setWriteToSecLog(b);
-		}
-
-		@Override
-		public boolean isWriteToSecLog() {
-			return owner.isWriteToSecLog();
-		}
-
-		@Override
-		public void setSecLogSessionKeys(String string) {
-			owner.setSecLogSessionKeys(string);
-		}
-
-		@Override
-		public String getSecLogSessionKeys() {
-			return owner.getSecLogSessionKeys();
-		}
-
-		@Override
-		public void registerEvent(String description) {
-			owner.registerEvent(description);
-		}
-
-		@Override
-		public void throwEvent(String event, Message eventMessage) {
-			owner.throwEvent(event, eventMessage);
-		}
-
-		@Override
-		public boolean sizeStatisticsEnabled() {
-			return owner.sizeStatisticsEnabled();
 		}
 	}
