+++ core/src/main/java/org/frankframework/processors/MonitoringPipeProcessor.java
@@ -15,6 +15,7 @@
 */
 package org.frankframework.processors;
 
+import io.micrometer.core.instrument.DistributionSummary;
 import jakarta.annotation.Nonnull;
 import jakarta.annotation.Nullable;
 import org.apache.commons.lang3.StringUtils;
@@ -25,12 +26,9 @@ import org.frankframework.core.PipeLineSession;
 import org.frankframework.core.PipeRunException;
 import org.frankframework.core.PipeRunResult;
 import org.frankframework.functional.ThrowingFunction;
-import org.frankframework.pipes.AbstractPipe;
 import org.frankframework.stream.Message;
 import org.frankframework.util.AppConstants;
 import org.frankframework.util.LogUtil;
-
-import io.micrometer.core.instrument.DistributionSummary;
 
 /**
  * @author Jaco de Groot
@@ -57,7 +55,7 @@ public class MonitoringPipeProcessor extends PipeProcessorBase {
 			summary.record(pipeDuration);
 
 			if (pipe.getDurationThreshold() >= 0 && pipeDuration > pipe.getDurationThreshold()) {
-				durationLog.info("Pipe [" + pipe.getName() + "] of [" + pipeLine.getOwner().getName() + "] duration [" + pipeDuration + "] ms exceeds max [" + pipe.getDurationThreshold() + "], message [" + message + "]");
+				durationLog.info("Pipe [{}] of [{}] duration [{}] ms exceeds max [{}], message [{}]", pipe::getName, () -> pipeLine.getOwner().getName(), () -> pipeDuration, pipe::getDurationThreshold, () -> message);
 				pipe.throwEvent(IPipe.LONG_DURATION_MONITORING_EVENT);
 			}
 		}
@@ -72,10 +70,8 @@ public class MonitoringPipeProcessor extends PipeProcessorBase {
 		sb.append("Pipeline of adapter [").append(ownerName).append("] messageId [").append(pipeLineSession.getMessageId()).append("] is about to call pipe [").append(pipe.getName()).append("]");
 
 		boolean lir = AppConstants.getInstance().getBoolean("log.logIntermediaryResults", false);
-		if (pipe instanceof AbstractPipe ap) {
-			if (StringUtils.isNotEmpty(ap.getLogIntermediaryResults())) {
-				lir = Boolean.parseBoolean(ap.getLogIntermediaryResults());
-			}
+		if (StringUtils.isNotEmpty(pipe.getLogIntermediaryResults())) {
+			lir = Boolean.parseBoolean(pipe.getLogIntermediaryResults());
 		}
 		if (lir) {
 			sb.append(" current result ").append(message == null ? "<null>" : "(" + message.getClass().getSimpleName() + ") [" + message + "]").append(" ");
